@@ -1,5 +1,5 @@
 {
-  "name": "anyagirlchild",
+  "name": "anyagirlchildngo",
   "version": "0.1.0",
   "private": true,
   "scripts": {
@@ -11,62 +11,54 @@
   },
   "dependencies": {
     "@ant-design/nextjs-registry": "^1.0.2",
-    "@prisma/client": "^6.3.1",
-    "@reduxjs/toolkit": "^2.5.1",
-    "@remirror/pm": "^3.0.1",
-    "@remirror/react": "^3.0.2",
-    "@remirror/react-editors": "^2.0.2",
-    "@tanstack/react-query": "^5.66.0",
-    "@tinymce/tinymce-react": "^6.1.0",
+    "@prisma/client": "^6.10.1",
+    "@reduxjs/toolkit": "^2.8.2",
+    "@tanstack/react-query": "^5.81.2",
     "@vercel/analytics": "^1.5.0",
+    "antd": "^5.26.1",
     "aos": "^2.3.4",
     "apexcharts": "^4.7.0",
-    "axios": "^1.7.9",
-    "bcryptjs": "^2.4.3",
-    "cloudinary": "^2.5.1",
+    "bcryptjs": "^3.0.2",
+    "cloudinary": "^2.7.0",
     "cookie": "^1.0.2",
     "dompurify": "^3.2.6",
     "handlebars": "^4.7.8",
-    "html-react-parser": "^5.2.2",
-    "html2pdf.js": "^0.10.2",
+    "html-react-parser": "^5.2.5",
     "jsonwebtoken": "^9.0.2",
     "lodash": "^4.17.21",
     "moment": "^2.30.1",
-    "next": "^14.2.29",
+    "next": "15.3.4",
     "next-auth": "^4.24.11",
     "next-cloudinary": "^6.16.0",
-    "nodemailer": "^6.10.0",
+    "nodemailer": "^7.0.3",
     "p-limit": "^6.2.0",
-    "react": "^18",
+    "react": "^19.0.0",
     "react-apexcharts": "^1.7.0",
     "react-country-region-selector": "^4.0.2",
-    "react-dom": "^18",
-    "react-hot-toast": "^2.5.1",
-    "react-icons": "^5.4.0",
+    "react-dom": "^19.0.0",
+    "react-icons": "^5.5.0",
     "react-paystack": "^6.0.0",
-    "react-phone-number-input": "^3.4.11",
     "react-redux": "^9.2.0",
-    "react-share": "^5.2.0",
-    "react-typing-effect": "^2.0.5",
-    "sharp": "^0.33.5",
-    "swiper": "^11.2.2",
+    "react-share": "^5.2.2",
+    "rich-text-editor": "^8.10.0",
+    "sharp": "^0.34.2",
+    "swiper": "^11.2.8",
     "swr": "^2.3.3",
-    "tinymce": "^7.9.0",
     "xlsx": "^0.18.5"
   },
   "devDependencies": {
+    "@eslint/eslintrc": "^3",
+    "@tailwindcss/postcss": "^4",
     "@types/aos": "^3.0.7",
-    "@types/bcryptjs": "^2.4.6",
-    "@types/jsonwebtoken": "^9.0.8",
-    "@types/node": "^20.17.17",
+    "@types/bcryptjs": "^3.0.0",
+    "@types/jsonwebtoken": "^9.0.10",
+    "@types/node": "^20",
     "@types/nodemailer": "^6.4.17",
-    "@types/react": "^18",
-    "@types/react-dom": "^18",
-    "@types/react-typing-effect": "^2.0.7",
-    "eslint": "^8",
-    "eslint-config-next": "14.2.5",
-    "postcss": "^8",
-    "prisma": "^6.3.1",
+    "@types/react": "^19",
+    "@types/react-dom": "^19",
+    "eslint": "^9",
+    "eslint-config-next": "15.3.4",
+    "prisma": "^6.10.1",
     "rimraf": "^6.0.1",
     "tailwindcss": "^3.4.1",
     "typescript": "^5"
