@@ -1,22 +1,36 @@
+import React from 'react'
 import { ASSET_URL } from '@/assets'
 import { PBLoginContainer } from '@/modules/public/pbauth'
 import { Metadata } from 'next';
 import Image from 'next/image'
-import React from 'react'
-
+import { config } from '@/config';
 
 export const metadata: Metadata = {
+  metadataBase: new URL(config.APP_PUBLIC_SITE_URL),
   title: "Anya Girlchild :: Login",
   description: "Anya Girlchild Foundation was born from a deep-seated personal experience with the challenges many young girls face in pursuing an education and achieving their full potential.",
+  icons: ASSET_URL["anya_girlchild_logo"].src,
+  openGraph: {
+    type: "website",
+    title: "Anya Girlchild :: Login",
+    images: [
+      { url: ASSET_URL["donation_man"].src, width: 800, height: 600 },
+      { url: ASSET_URL["donation_man"].src, width: 1800, height: 1600 },
+    ],
+    siteName: "Anya Girlchild Foundation",
+    description: "Anya Girlchild Foundation was born from a deep-seated personal experience with the challenges many young girls face in pursuing an education and achieving their full potential.",
+    locale: 'en_US',
+  }
 };
 
 type TPageProps = {
   searchParams: {
-    view?: string
+    view?: string | null
   }
 }
 
-export default function PBLoginPage({ searchParams: { view } }: TPageProps) {
+export default async function PBLoginPage({ searchParams }: TPageProps) {
+  const { view } = await searchParams;
   return (
     <main className='flex flex-col md:flex-row gap-4 lg:gap-8 md:h-full'>
       <section className="container mx-auto flex flex-col lg:flex-row gap-4 md:h-full">
