@@ -2,19 +2,20 @@ import type { Metadata } from "next";
 import { Poppins, Space_Grotesk } from 'next/font/google'
 import "../globals.css";
 import { Analytics } from '@vercel/analytics/next';
-// import { AosProvider, Footer, Header } from '@/components'
 
 import { AntdRegistry } from '@ant-design/nextjs-registry'
 import ThemeProvider from "@/provider/ThemeProvider";
-import { Toaster } from "react-hot-toast";
 import PBHeader from "@/modules/public/pblayout/PBHeader";
 import PBFooter from "@/modules/public/pblayout/PBFooter";
 import { AosProvider } from "@/modules/shared";
+import { config } from "@/config";
 
 const grotesk = Space_Grotesk({ subsets: ["latin"], variable: "--grotesk", weight: ["300", "400", "500", "600", "700"], fallback: ["cursive"] });
 const poppins = Poppins({ subsets: ["latin"], variable: "--poppins", weight: ["100", "200", "300", "400", "500", "600", "700", "800", "900"], fallback: ["cursive"] });
 
+
 export const metadata: Metadata = {
+  metadataBase: new URL(config.APP_PUBLIC_SITE_URL),
   title: "Anya Girlchild :: Home",
   description: "Anya Girlchild Foundation was born from a deep-seated personal experience with the challenges many young girls face in pursuing an education and achieving their full potential.",
 };
@@ -27,10 +28,9 @@ export default function RootLayout({
 
   return (
     <html lang="en">
-      <body className={`${grotesk.variable} ${poppins.variable} antialiased relative`}>
+      <body className={`${grotesk.variable} ${poppins.variable} font-poppins antialiased relative`}>
         <AntdRegistry>
           <ThemeProvider>
-            <Toaster />
             <AosProvider>
               <PBHeader />
               <div className="font-poppins min-h-[80vh]"> {children} </div>
