@@ -15,9 +15,13 @@ export const metadata: Metadata = {
   openGraph: {
     type: "website",
     title: "Anya Girlchild :: Partner",
-    images: ASSET_URL["little_child"].src,
-    siteName: "Anya Girlchild Foundation",
-    description: "Anya Girlchild Foundation was born from a deep-seated personal experience with the challenges many young girls face in pursuing an education and achieving their full potential.",
+    images: [
+          { url: ASSET_URL["hungry_man"].src, width: 800, height: 600 },
+          { url: ASSET_URL["hungry_man"].src, width: 1800, height: 1600 },
+        ],
+        siteName: "Anya Girlchild Foundation",
+        description: "Anya Girlchild Foundation was born from a deep-seated personal experience with the challenges many young girls face in pursuing an education and achieving their full potential.",
+        locale: 'en_US',
   }
 };
 
